lingvo/requirements.txt
Django==1.9.7
Markdown==2.6.6
Pillow==3.2.0
Twisted==16.2.0
asgi-redis==0.13.1
asgiref==0.13.3
autobahn==0.14.1
channels==0.14.1
daphne==0.12.1
django-allauth==0.25.2
django-filter==0.13.0
django-rest-auth==0.7.0
djangorestframework==3.3.3
djangorestframework-gis==0.10.1
msgpack-python==0.4.7
oauthlib==1.1.1
psycopg2==2.6.1
python-openid==2.2.5
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
txaio==2.5.1
wsgiref==0.1.2
zope.interface==4.2.0